+++ src/components/Node.cpp
@@ -7,7 +7,7 @@
 #include "Node.h"
 #include "NodeImpl.h"
 
-Node::Node(std::string _name, NODE_TYPE _nodeType, uint32_t _dpr) {
+Node::Node(std::string _name, NODE_TYPE _nodeType, uint64_t _dpr) {
     uuid_generate(this->id);
     this->name = std::move(_name);
     this->nodeType = _nodeType;
@@ -45,7 +45,7 @@ size_t Node::nodeToIdx(const std::map<size_t, Node *> &map, const Node *node) {
  * @brief get precess delay of node, unit： ns
  * @return   uint32_t
  * */
-uint32_t Node::getDpr() const {
+uint64_t Node::getDpr() const {
     return dpr;
 }
 
@@ -53,7 +53,7 @@ uint32_t Node::getDpr() const {
  * @brief set precess delay of node.
  * @param   _dpr    process delay, unit： ns
  * */
-void Node::setDpr(uint32_t _dpr) {
+void Node::setDpr(uint64_t _dpr) {
     Node::dpr = _dpr;
 }
 
@@ -64,7 +64,7 @@ void Node::setDpr(uint32_t _dpr) {
  * @param   dpr         node process delay, unit: ns
  * @return  Node *
  * */
-Node *createNode(NODE_TYPE nodeType, const std::string& name, uint32_t dpr){
+Node *createNode(NODE_TYPE nodeType, const std::string& name, uint64_t dpr){
     Node *node;
     switch (nodeType) {
         case END_SYSTEM:
